postManagementService/requirements.txt
dnspython==2.3.0
grpcio==1.30.0
grpcio-tools==1.30.0
protobuf==3.20.0
pydantic==1.10.9
pymongo==4.4.0
python-dotenv==1.0.0
typing_extensions==4.6.3
colorlog==6.7.0
schedule==1.2.0

